Rename TestModuleRelayer to TestModule
